requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
tensorflow>=2.13.0

# Visualization
matplotlib>=3.7.0

# API & WebSocket
pyotp>=2.9.0
smartapi-python>=1.3.0

# Logging (if using logzero in ws_adapter.py)
logzero>=1.7.0
